code cleanup using custom hooks
