+++ src/pages/auth/login.jsx
@@ -1,44 +1,22 @@
-import { Link, useNavigate, useLocation } from "react-router-dom";
+import { Link } from "react-router-dom";
 import "styles/auth/login.css";
 import { useEffect } from "react";
-import axios from "axios";
-import { useState } from "react";
-import { useAuth } from "contexts";
+import { useLogin } from "utils";
 
 export const Login = () => {
-  const [_email, setEmail] = useState("");
-  const [_password, setPassword] = useState("");
-  const { setIsLoggedIn } = useAuth();
-  const navigate = useNavigate();
-  const location = useLocation();
+  const {
+    loginHandler,
+    dummyHandler,
+    _email,
+    _password,
+    setEmail,
+    setPassword,
+  } = useLogin();
 
   useEffect(() => {
     document.title = "Login | Laugh Factory";
   }, []);
 
-  const loginHandler = async (event) => {
-    event.preventDefault();
-    try {
-      const response = await axios.post("/api/auth/login", {
-        email: _email,
-        password: _password,
-      });
-      console.log(response);
-      localStorage.setItem("token", response.data.encodedToken);
-      setIsLoggedIn(true);
-      navigate(location.state?.from?.pathname || "/videoListing", {
-        replace: true,
-      });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
-  const dummyHandler = (event) => {
-    event.preventDefault();
-    setEmail("arpitkumar@gmail.com");
-    setPassword("arpit1234");
-  };
   return (
     <section className="loginContainer">
       <div className="heading">
