+++ src/pages/auth/signup.jsx
@@ -1,63 +1,15 @@
-import { Link, useLocation, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import "styles/auth/signup.css";
-import { useEffect, useReducer } from "react";
-import axios from "axios";
-import { useAuth } from "contexts";
+import { useEffect } from "react";
+import { useSignup } from "utils";
+
 
 export const Signup = () => {
-  const { setIsLoggedIn } = useAuth();
-  const location = useLocation();
-  const navigate = useNavigate();
+  const { submitHandler, signUpDispatch, signUpState } = useSignup();
 
   useEffect(() => {
     document.title = "Signup | Shopzila";
   }, []);
-
-  const signUpReducer = (signUpState, signUpAction) => {
-    switch (signUpAction.type) {
-      case "FIRSTNAME":
-        return { ...signUpState, firstName: signUpAction.payload.target.value };
-      case "LASTNAME":
-        return { ...signUpState, lastName: signUpAction.payload.target.value };
-      case "EMAIL":
-        return { ...signUpState, email: signUpAction.payload.target.value };
-      case "PASSWORD":
-        return { ...signUpState, password: signUpAction.payload.target.value };
-      case "CONFIRM_PASSWORD":
-        return {
-          ...signUpState,
-          confirmPassword: signUpAction.payload.target.value,
-        };
-      default:
-        return signUpState;
-    }
-  };
-
-  const [signUpState, signUpDispatch] = useReducer(signUpReducer, {
-    firstName: "",
-    lastName: "",
-    email: "",
-    password: "",
-    confirmPassword: "",
-  });
-
-  const submitHandler = async (event) => {
-    event.preventDefault();
-    try {
-      const response = await axios.post(`/api/auth/signup`, {
-        firstName: signUpState.firstName,
-        lastName: signUpState.lastName,
-        email: signUpState.email,
-        password: signUpState.password,
-        confirmPassword: signUpState.confirmPassword,
-      });
-      localStorage.setItem("token", response.data.encodedToken);
-      setIsLoggedIn(true);
-      navigate(location.state?.from?.pathname || "/videoListing", { replace: true });
-    } catch (e) {
-      console.log(e);
-    }
-  };
 
   return (
     <section className="signupContainer">
