+++ src/reducer/index.js
@@ -3,6 +3,7 @@ import { watchLaterReducer } from "./watch-later";
 import { likedVideoReducer } from "./liked-videos";
 import { historyReducer } from "./history";
 import { playlistReducer } from "./playlist";
+import {signUpReducer} from "./signup";
 
 export {
   videoReducer,
@@ -10,4 +11,5 @@ export {
   likedVideoReducer,
   historyReducer,
   playlistReducer,
+  signUpReducer
 };
