+++ src/utils/login.jsx
@@ -1,0 +1,44 @@
+import { useNavigate, useLocation } from "react-router-dom";
+import axios from "axios";
+import { useState } from "react";
+import { useAuth } from "contexts";
+
+export const useLogin = () => {
+  const [_email, setEmail] = useState("");
+  const [_password, setPassword] = useState("");
+  const { setIsLoggedIn } = useAuth();
+  const navigate = useNavigate();
+  const location = useLocation();
+
+  const loginHandler = async (event) => {
+    event.preventDefault();
+    try {
+      const response = await axios.post("/api/auth/login", {
+        email: _email,
+        password: _password,
+      });
+      console.log(response);
+      localStorage.setItem("token", response.data.encodedToken);
+      setIsLoggedIn(true);
+      navigate(location.state?.from?.pathname || "/videoListing", {
+        replace: true,
+      });
+    } catch (e) {
+      console.log(e);
+    }
+  };
+
+  const dummyHandler = (event) => {
+    event.preventDefault();
+    setEmail("arpitkumar@gmail.com");
+    setPassword("arpit1234");
+  };
+  return {
+    loginHandler,
+    dummyHandler,
+    _email,
+    _password,
+    setEmail,
+    setPassword,
+  };
+};
